chore: create resolveProcessCredentials
